// frontend/src/core/api/generated/model/siteAttendanceEntries.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { AttendanceEntryRequest } from './attendanceEntryRequest'

export interface SiteAttendanceEntries {
  siteId: string
  /** @minItems 1 */
  entries: AttendanceEntryRequest[]
}