// frontend/src/core/api/generated/model/siteAttendanceSheet.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Kızılkan Şantiye API
 * OpenAPI spec version: 1.0
 */
import type { AttendanceDayView } from './attendanceDayView'
import type { WorkerView } from './workerView'

export interface SiteAttendanceSheet {
  siteId: string
  siteName: string
  workers: WorkerView[]
  day: AttendanceDayView
}